Remove "Clowder is not enabled" message
Hello, in our app, we use the same configuration for the backend app as well as for many CLI tools and this message is just always there. I see no big value in having anything on standard output, we use logging (and cloudwatch) anyways.

I'd appreciate a minor release with this so it goes away for good :-) Thanks a bunch, cheers!